.history/requirements_20260314201149.txt
streamlit
google-genai
PyPDF2
python-docx
python-dotenv